Move bq code to seperate module
